4214: Implement theme hook for Ting titles
This hook can be reused across standard field based displays as well
as fast rendering as long as a collection is provided. Other
variables can be extracted from the field settings if available.

This provides a better separation than the current function per field
approach in ting_search_fast as it allows better reuse across
implementations.

The material type is a part of the title field so we can remove this
from the fast search implementation.