4261: Use change event instead of click on select all renew
Something has changed so that using a click handler for these buttons
breaks the functionality of the select all renew checkbox and therefor
also the global select all renew button (which just reuses the select
all checkbox).

The problem was the triggering of click handler on the local select all
checkboxes, from the change event handler on the global select all. This
would simuluate a real click chaning state of the checkbox and mess up
the logic in the click handler.

The easiest fix seem to be to use a change handler instead on the local
select all checkboxes. This will not break any functionality since a
real click from a user will also fire the change handler. And now we can
trigger the change handler instead, which doesn't change the state of
the checkbox. It also seems more consistent since a change handler is
alos used on the other checkboxes.