4207: Upgrade EU Cookie Compliance module to release 1.28
This addresses a cross site scripting security issue:
https://www.drupal.org/sa-contrib-2019-033

The update is a major leap in versions where the functionality of the
module has been extended to take GDPR and other issues into account.

We do not have to do anything special regarding consent here so we
just need to configure the module to behave as it did previously.
With the new configuration options this boils down to:

1. Using the default consent method
2. Not showing any disagreement button

The updated module uses I18n variable to store internationalized
settings. To provide a consistent configuration we also use this
module for new installs. Instead of defining all configuration
options we modify the defaults to work as we want to. This allows us
to highlight what is important in our case.

Our custom template has been updated to both include the close button
and be in line with the latest changes to the module.

Finally we have to explicitly specify paragraphs in our cookie text
to be in block format. For some reason the updated module thinks
they should be inline.