4207: Fix close button for EU Cookie Compliance popup
In the current form the button does nothing - probably because of a
race condition. The behavior was called before the popup was created.

To fix this we instead listen for an event emitted by the module and
register our click handler then.

Click logic is replicated from 
Drupal.eu_cookie_compliance.declineAction to ensure that the close
animation is similar to the open animation. Before it was too fast.