4214: Implemented support for rating display without entities
The implementation is placed within the ding_entity_rating module
despite the fact that it does not use entities. The fact that the
module deals with ratings makes it the most relevant location.

To support retrieval of ratings without an entity we also update the
key function for entity rating retrieval: ding_entity_rating_average.
Instead of accepting an entity it now only requires an id. The
ding_entity_id and the Ting object id are the same value. The only 
calling functions are accessors in the same module so it should be
safe to update.

Update fast template to mimic the default output of the rating field.