4295: Hide ratings on search results
The current implementation loads ratings synchronously for each
result. This increases response times considerably.

To address this we hide the ratings field on search results. This is
also in line with UX recommendations.

If we want to reintroduce ratings they should be loaded
asynchronously instead.

This change is the result of hiding the ratings field group and
fields on the search result view mode for ting objects and doing a
features export.