4319: Use FAPI #required property instead of custom
The motivation for this change was to get the required asterisk on the
required p2b fields. The best way to get this working is to remove our
custom validation and use FAPI #required property. Since these are shown
at labels, we also need to add a title to these fields.