4322: Do not display async empty campaign plus pane wrapper
The default panel pane markup surrounding campaign plus panel panes
has styling. It will be rendered when using asynchronous loading
but only the content of the pane is replaced with actual content. If 
there are no campaigns to be shown in the panel pane this means that 
the surrounding markup remains. As it is styled with a padding and 
background color this leaves an empty grey box.

To avoid this we hide the entire wrapper by default and only show
it if any actual campaign markup has been loaded.