Core: Find covers using CSS instead of XPath in Behat tests
For some reason cover images are no longer identified in search
results when using XPath. However it works when using CSS selectors.

One reason might be the addition of new alt attributes and titles on
links and images causing trouble with special characters or the like.
On the other hand tests worked when they were merged.