2697: Make list title errors in Behat tests easier to read
Currently tests will fail with a message that two boolean values are
not equal. You do not get any information that this occurs in
relation to the list title.

To address this we update the list page object to return the list
title and compare that string to the expected title.