4379: Fix progressing through search result for carousel covers
The current implementation accepts a start offset and returns
the offset that a subsequent request should start at. Currently the 
same value, $start, is returned as is passed without being changed.

This will result in the same request being repeated again and again
if sufficient covers are not found. The same request will not yield
additional covers and thus this results in request timeouts.

To address this we return the passed offset + the number of search
results to retrieve in the current request. This ensures that a
subsequent request retrieves a new chuck of search results.