1460: Use provider function for checking reservability
Using provider function instead of ding entity pseudo classes seems to
be the most practical way to be able to check sevaral at once.

It also makes sense that the reservation provider is solely responsible
for checking reservability, and that it's not something that every
module can hook into and have a say in.