4467: Add single signout endpoint for adgangsplatformen
This adds a new menu callback which can be used as a single signout
with Adgangsplatformen.

The endpoint will be embedded in an iframe on Adgangsplatformen when
using single signout. Consequently we have to take care how status
is reported.

See https://github.com/DBCDK/hejmdal/blob/master/docs/single-logout.md