4126: Remove not needed cache_key_alter implementation
After the work with SAL the timing of the the generation of the cache
key has changed, such this implementation is not needed anymore for
opensearch to be able to serve the correct response when using profiles.

Before hook_opensearch_pre_execute was invoked after generation of cache
key and as a result it didn't take into account any modification made in
pre_execute implementation. This couldn't be changed without moving a
lot of code around, so the easiest solution at that time was to
introduce a new cache_key hook, such that ting_field_search could modify
it based on an active profile.

Hence in this commit we remove ting_field_search cache_key_alter
implementation. Since it's not needed it's better to remove to reduce
complexity.