3021: Prevent errors when migrating nodes with workflows
When content is migrated to Paragraphs this also causes the host
node to be saved. If the node has workflows then these are loaded.
However an update hook usually runs as the anonymous user which does
not have permissions to perform any workflow transitions. This leads
to an error: "There are no workflow states available".

To fix this we run the migration as user 1 instead. This user is
allowed to perform all transitions.

No actual transitions are performed but this removes the error
nonetheless.