4394: Only add original price if price is existing and has id
We should only add reference to original price if the current price is
existing and have a price id in p2b. Otherwise our sync function will
try to update a non-existing price.

Note that we need to check for $price->id. It's not enough to verify
price is non-empty since it can be a clone which has data but no id.