4381: Disable Autologout once Adgangsplatformen is installed
Autologout in combination with single signout from Adgangsplatformen
will log out the user from all platforms when inactive on one. That
is not helpful - especially in combination with platforms used for
passive media consumption.

We do not remove the module entirely as some libraries may want to
retain the autologout functionality. They will have to reenable it
themselves subsequently.