4652: Add support for memcache
We use this in production and some development environments so we
might as well use it here. It might also make the site run faster.

This follows the installation guide from Platform.sh although we have
to modify to path to the memcache files because we use an
installation profile.