4514: Improve display of login form on /user
In the current state form items are floated left. There does not
seem to be a reason for this. They are already in a container which
uses floats for positioning and they are positioned directly above/
below each other.

Also add a max-width for the form actions. This ensures that the
login button will not be longer that form items which currently
also have a 400px max-width.

The login form in the popup bar already have more specific styling
which overrides the newly added max-width value.