4520: Prevent JavaScript errors with ding_popup
AJAX commands in Ding Popup requires the existence of the Drupal AJAX
core library. This is not handled by the current definition in the
.info file.

Such an error can block the execution of other JavaScript code e.g.
the handling of the top menu.

To address this we declare the code a library and with the AJAX
library as a dependency.

We load the library is currently loaded for all HTML responses as we
cannot establish exactly where popups are invoked.