4520: Prevent JavaScript errors with ding_serendipity
AJAX commands in Ding Popup requires the existence of the Drupal AJAX
core library. This is not handled by the current definition in the
.info file.

Such an error can block the execution of other JavaScript code e.g.
the handling of the top menu.

By searching through the codebase it seems as if the AJAX command 
defined in the code is not in use. Consequently we remove it.