SEO: Update ting_object book schema_metatag config
.. to support multiple creators and contributers using new token
and pivot option.

Note that the trailing comma is intended as this is the only way
to make it work properly. With out this no person entity will be
outputted if there's only one creator/contributer. The pivot
option is build to support comma-separated tokens in the config
and not a token that returns mutliple comma-separated values. So
the work around is the insert the comma after our token and it
seems to work in both cases.

This might be a bit hacky but it will work for now. The
alternative is to create an issue and patch, but this might be a
complicated matter.