4555: Move teaser ding_entity buttons to configuration
Before they were added in the theme probably because only some of
the buttons is wanted in teasers and not all returned by the
ding_entity_buttons hook. Because of this Ding Webtrekk was
failing to attach event in hook_entity_view() since this is
invoked way before theme processors in the theme layer.

To solve this we correctly add the ding_entity buttons in field
config for the teaser views modes. Any unwanted stuff must then be
hidden with styling or removed/moved in theme processors.