4604: Improve “Closed” text for opening hours without categories
If a library is not using categories for opening hours then the
current codebase will still load all available categories and show
these as closed despite opening hours data being available without
categories.

Consequently we change the code from creating “closed” texts without
categories and then trying to massage the texts into the categories
if needed to actively generating closed texts for all categories
that are in used for the current instances.