4612: Improve redirection to /user/me from /user/uid in urls
This prevents uids from being picked up in trackers which can lead
to linking of user identity to specific user behavior.

Before this change users could access urls in the /user/uid format.
If not logged in then this url would display “page not found”. If the
user was logged in then this url would work just as well as the
preferred /user/me format.

To avoid these urls from being tracked we redirect immediately to
the frontpage for user pages without access and for users seeing
their own profile then redirect to /user/me.

Uids will still show up in urls where an administrative user accesses
another user but that is acceptable.