Set token on page build instead of user login
User login may be invoked as a part of a redirect scheme and 
consequently no JavaScript will be executed and no value stored in
local storage.

We run it at page build instead. This will occur multiple times but
there is little harm in setting the same value more than once.