Scramble single signon names when using Connie
The current implementation takes the username directly and uses that
for display and email. However when using Adgangsplatformen the name
will actually be the CPR number.

We should not display such information unnecessarily so scramble it
using a CRC32 - a hash function which will generate a reasonable
short and probably unique value suitable for display.