Add timeout handling for user list migration
We want a low timeout to prevent the system from stalling during
login - especially if the list services perform poorly. This effect
will multiply if many users wait for a long time for migration to
complete. We prefer a lower timeout which in the case of slowness
will migrate a user the next time they log in.

5 seconds seems like a suitable number. We do not expect the
migration to take more that 1 second but in the case of request
pileup we do not want them to wait any longer than that.