Load React component user data in separate non-cacheable request
Ding with its Varnish configuration will cache requests even for
authenticated users. This means that we cannot rely on the current
approach where we render the token as a part of the DOM as this will
be cached and cause user tokens to get mixed up.

Instead we setup the token in a separate JavaScript file which is
loaded in a separate request from other JavaScript resources.

This is supported by the recent update to the DDB React project which
supports in-memory token handling.