Reintroduce user profile styling
During the P2 exodus it seems as if a bit too much was removed. The
current styling of /user/me/view is a bit lackluster. It can be 
accessed when clicking directly on the my pages button. The page 
uses a pane from ding_user_frontend which it turns out uses styling 
that previously decided within the now removed 
themes/ddbasic/sass/p2/p2-class-panel.scss file.

To address this we add a new panel SCSS file containing the required
parts of the styling from the file and add that to the theme.