4688: Sort search results from saved searches by date descending
We do not have a way to identify new materials at the moment. Thus
it is reasonable to assume that the user wants to see the most recent
materials first.

We do not have a correct acquisition date index so we call back on
the publication date instead.

Since the sorting is a query parameter which we control ourselves we
can just update the string template to add the sort parameter.