4688: Use provider function if available to generate query to follow
By using the provider function we allow the provider to transform
the entire query object to a string. This includes things like
filters (facets).

Since CQL doctor usage is part of the OpenSearch implementation we
do not have to call that explicitly either.

To avoid making it a required provider hook we fall back on the 
full text query part of the search request object.