Migrate all ids for the user on Adgangsplatformen login
In the current situation we may have different legacy OpenList user
ids for a user depending on whether they logged in using CPR number
or borrower card number. This results in them currently having
different checklists. 

Going forward Adgangsplatformen, MaterialList and FollowSearches will 
alleviate this by using a single GUID for the user which is separate
from the credential types used. 

During the migration we have a chance to remedy this split-up by
migrating data for the different legacy user ids to the same GUID.

To implement this we rely on Adgangsplatformens ability to provide
all the different user ids available for a user for an agency/site. 
We can then map these to corresponding Drupal user accounts on that 
site which may have different OpenList user ids.

By migrating all of the accounts using a token belonging to the same 
user then all data belonging to each account will be migrated to the
same user. 

We have to perform the migration during the Adgangsplatformen 
callback after the user has logged in as this is the only context 
where we have the necessary information available. We descale the
information available through the session token to avoid this
information from leaking. Thus we cannot perform this operation in
ding_react_user_login() or the like.