Use multi insert when importing KPI data during cron
This will greatly improve performance with MYSQL backends. Several
orders of magnitude.

The number of rows we can insert each in each query depends on the
max_allowed_packet setting in MYSQL. We default to 10.000 which
already improves performance considerable and should be safe with
the standard value of max_allowed_pachet of 16M. The config can
easily be changed so we provide a settings.php config to control
chunk size for each insert:

'ting_subsearch_kpi_multi_insert_chunk_size'