Bypass Varnish cookie stripping for React user token javascript
Our current configuration in ding2.vcl will strip cookies from as
requests ending in .js assuming they are static files which should
always be cacheable. 

Our user token callback currently is user.js and is consequently also
affected by this. This results in the session cookie being stripped
and thus the callback will always reply as if the user is not logged
in.

Rename the file path of callback from user.js to user-js as that 
means the response will not be treated as a static file.