4752: Ensure AJAX responder is loaded when using Adgangsplatformen
We rely on redirect responses which will pass the user along to
Adgangsplatformen if she tries to perform an action while not logged
in.

If the CTools AJAX responder is not loaded then the responder will be
returned and loaded as a part of the response. This has two problems:

1. The user is not redirected to Adngagsplatformen immediately but
must try to perform the action again.
2. The user is not redirected to the correct path e.g. when trying to
make a reservation.

Being aware of this we originally intended to include the AJAX
responder with every anonymous response but a refactor broke the
hook name which should ensure this.

Rename the function ding_page_build() to 
ding_adgangsplatformen_page_build() to implement the hook properly.

Since the current code is already running in production I assume we 
have been saved by another part of the project including the file
which has subsequently been disabled/removed. Probably in the P2
modulespace.