4756: Use Adgangsplatformen authentication client when logging out
We currently use a separate secondary Adgangsplaformen client for
retrieving tokens without access to sensitive information. These 
tokens are suitable for use in client side applications.

However the secondary client is potentially shareable between 
installations and thus cannot be configured with a single valid 
logout url. This means that Adgangsplatform cannot redirect to the
correct site.

To address this we instead revoke the access token manually and then
generare a unique access token for an anonymous user using the
primary client.

We can then pass this token to Adgangsplatformen when logging out and
it will allow redirect to the corresponding site.