4761: Never cache ding_react/user.js in Varnish
This file contains a unique per user specific token and should not
be cached. However our current Varnish configuration treats all files
ending in .js as a static file and strips cookies no matter what
headers we send.
 
This results in the file always returning as if accessed by the
anonymous user.

To fix this we update the Varnish configuration to bypass caching of
this file.