4614: Library proivder decides whether to show periodical holdings
Instead of using hook_entity_is periodical pseudo-class to decide
whether to show periodical holdings, and basing the decision on
bibliographic data, we know base the decision on the data returned
from the library provider. If there's periodical issues in the
holdings data, we show the periodical holdings.

We thereby give the library system full control over when to show
periodical holdings.

Note that we still use hook_entity_is and only bother to show the
field if the entity is considered library material and have a
chance to have periodical holdings.