4765: Use jQuery.attr() instead of data() in reservability checks
jQuery.data() will try to process the data value to an appropriate
type and this “00123” becomes 123.

That is not desirable for us and thus we fall back to using the
simpler jQuery.attr() function.

“Every attempt is made to convert the attribute's string value
to a JavaScript value (this includes booleans, numbers, objects,
arrays, and null) […] To retrieve a data-* attribute value as an
unconverted string, use the attr() method.”
https://api.jquery.com/data/#data-html5