Core: Reorder Ding2.install update hooks
The deployed version of the installation contains a patch where the
update hook ding2_update_7075 is defined and containing the code to
enable ding_registration.

This means that the current version of ding2_update_7075() will not
be executed. This is a problem as it now contains a different change.

To maintain the order and ensure that all necessary update hooks are
executed we must redefine ding2_update_7075() to contain the 
expected change and renumber the update hooks so they follow after
ding2_update_7075().