3659: Ensure user cookies are removed
The contact module sets a cookie with the enteres name and mail
for anonymous users. The purpose is to remember what the user
entered and then autofill the name and mail fields on the contact
form.

This functionality doesn't even work though since misc/form.js is
never included, which contains the javascript logic to handle it.

In any case we should not have cookies laying around on devices
with name and mail information entered in the contact form.
Especially when they do not serve any purpose.

So here we do our best to ensure they are removed from the users
devices when they submit the form.