3585: Track opt-out event even in opt-in model
In opt-in model the user is in opt-out by default, but we will
first send an opt-out event when the user declines the cookies
pop-up. Then we may miss some users who ignore the pop-up
completely, but it's the best we can do. It doesn't make sense to
send pop-out events repeately on page-load until the user has made
a decision.