3583: Move session expire settings to ding_provider
In #1329 with the integration of adgangsplatformen, the Ding user
settings form was moved to ding_user_form.

This was overriding the expire settings added in this PR, which
was added to old the form.

Since ding_user module is no longer dealing with Drupal login,
which are now handled in ding_user_form and ding_adgangsplatformen
depending on login method, and the session expire functionality
should apply regardless of login method, we now move session
expire settings to ding_provider module on the general settings
form.

There's already a setting for deleting inactive provider users on
that form, so perpahps it makes sense to place it here.