4622: Only show "Other formats" button on entity path
The $view_mode can never be 'teaser', and will always be either
'default' or 'ajax', therefore the collection was actually always
loaded. See: ding_entity_field_formatter_view().

It's not possible to get the $view_mode of the entity being
viewed from a field formatter. This is by design by the looks of
it; the formatter should behave the same regardless of the view
mode if the entity it's being displayed on.

Since it make sense to only show the "Other formats" button on
the main page of the material, a better solution is to check for
the entity path.