3585: Ensure cookie compliance module is configured properly
The values we set in the default configuration must include all values
expected when reading the configuration again. The module will not
provide any sane defaults. This may result in warnings as well as
errors.

If the cookie_name variable is not set then we will not be able to
migrate cookie consent from an anonymous user to the authenticated
user on login.