3585: Disable Platform.sh HTTP cache
When it is enabled it will strip all cookies except the listed ones -
even for POST requests despite claimed otherwise by the
documentation.

This prevents the cookie compliance consent cookies from being 
detected and passed to the user by the storage provider during login.

We disable the cache entirely for now to be able to test this. If we
want HTTP cache in these environments at some point it should be
handled by Varnish using the configuration included with the code
base to improve parity with production environments.