4951: Perform 0-hit redirect in hook_ting_search_results_prefix()
This removes the old zero hit redirect plugin and performs the
redirect in a better timed hook instead.

Also switch to use drupal_set_message() to display message about
redirect to the user. The message is not a suggestion but more an
informal message to the user and drupal_set_message() is therefore
more appropiate. We can also get away with the "original-search"
query parameter.

This solution now uses session cookie instead which will bypass
Varnish. But the search should be cached and it will only be for
redirects, so it should not be a big deal.