4975: Wait up to 5 seconds when checking for object covers.
We see multiple failures where a cover cannot be found even though
the subsequent screenshot shows the cover as expected. However Covers
may be loaded asynchronously and may not yet be ready when the test
is conducted.

Consequently we wait a bit. This is the same approach taken on the
search page.