4969: Remove reservation and periodical HTML IDs
These are not used anywhere and should not be needed.

For periodicals with many issues it can result in a lot of IDs in
the holdings table, and all these are submitted back in AJAX calls
on the page and can in rare cases result in errors on periodical
reservation.