5068: Fix bibdk subsearch
This PR fixes two issues with the bibdk subsearch mechanism:

1. Since opensearch tries to predict cache by putting collections
from the search result in cache, the bibdk subsearch would result
in full bibdk collections later being shown to the user. As a
solution we implement hook_opensearch_cache_key_alter() and alter
the cache key if bibdk subsearch is currently being performed.

2. We currently turn off bibdk subsearch immediatly after the
subsearch is performed. This means that collections will be
fetched without bibdk agency and profile. This actually relied
on bug 1. to work, but when it's fixed the bug manifests by
displaying empty collections. The solution is turn off bibdk
subsearch after we have fetched collections.

Also refector the code that controls the static variables into
functions to make it cleaner.