3437: Do not assume all entities for a type is online
When creating render array for the availability types field
formatter we currently assume that all entities for a type
is online if we find that the first one is.

This is not correct as we can't know if online accessibility is
purely type based. It's up to how the search provider chooses to
implement isOnline() method of TingObjectInterface (see #1750
where this is fixed to work as intended - closely related to
these changes).

For example for opensearch it will check for online access URL
in dc:terms URI.