5073: Make title field required
There is a small issue with this where it can make the "Add one
more" AJAX button fail silently because of required validation
errors.

But since we fill the field with "%title" ctools token, the chance
of this happening is minimal and it's therefore preffered to
ensure that there will always be a title.