5106: Remove duplicate translation of string to "Gem"
Drupal uses button texts to determine which button was clicked.
This leads to problems when multiple buttons have the same text.
Drupal cannot figure out which button was clicked and will choose 
whatever.

This leads to problems which can be experienced on the Webform
components page. Here you cannot add more components than one as it
contains both an Add and a Save button. Both are translated to Gem.
Consequently Drupal always acts as if the Save button was clicked.

Actually the translation of Add is a duplicate so we just remove it.

The problem only seems to occur for newly installed sites so there
is no reason to run an update hook.