1663: Implement static cache for p2b events and prices
With date repeat entity we risk loading the same event and prices
multiple time during a request. At the same time we already did
this a few times during other hooks.

So to improve performance we implement a static cache for p2b
events and prices to keep the request to API to a minimum.