4276: Prevent overlap between teaser and button in nodelist + blocks
You can use nodelist to display nodes in blocks. Each node has a
teaser text and read more link in the form of a button. This is shown
when the user hovers node.

However when the truncated teaser spans the entire last line then the
read more button will overlap with the teaser text.

To prevent this we increase the size of the teaser when hovered. This
provides more space for the button which will no longer overlap with
the text.