4276: Hide lead text for e-resources in node blocks by default
Initially lead text in node blocks should be hidden and shown on 
hover. Currently this is achieved by setting the element-hidden class
on the text in the template for all other content types.

It seems to have been forgotten for e-resources so add it.

One might argue that such behavior should be managed entirely through
JavaScript but this is the approach taken by the module.