4276: Prevent button from overlapping lead text in nodelists
Currently sizing of nodes shown in novelists with and without hover
is handled through combinations of JavaScript and CSS handling
Height, min-height and max-height with and without hover. The
situation is further complicated through the possibility of 1-2 lines
of header and 1-3 lines of lead text. This can lead to many 
variations in actual size. Meanwhile we want to achieve a uniform 
design with when nodes are shown side by side three on a line.  

This can lead to situations where the read-more button is overlapping
the lead text.

Fix this by increase the bottom padding for pages and e-resources on
hover. The read-more button is positioned in the lower right corner
of the box.

This is a compromise between trying to fix the issue and reworking 
the entire handling. It should prevent overlap but sizing may not
appear optimal in all situations.