1663: Rework ding_place2book_get_event()
It's not a good idea to return the prices as prop of the event
object, since p2b API will actually try to create these as new
prices (regardless of existing id) of the event with prices prop
is sent.

So rework to return array with event object as first entry and
prices as second.

Also change the return value always return array and throw
exception (instead of returning FALSE) if something went wrong.
We still catch the exception to log generic watchdog message, but
re-throw right after.

This way the caller will have access to exception and more info
about what went wrong.