5133: Map readable FBS dept type to Ding dept type
Reason message is a free text field where library employees can enter
whatever. This does not go well with the design. Instead use the type
field.

We expect the type field to be restricted regarding allowed values
but official documentation of actual values is nowhere to be found.
Consequently we reference types which can be found in examples like
https://platform.dandigbib.org/issues/5059 and fall back to a generic
“other” for the rest.