5136: Improve layout of material item information elements
Currently we rely on dynamic resizing based on content. This results
in an uneven rhythm where you are not guaranteed that the same info
aligns vertically across items in a list.

Introduce a shared grid where each info has the same width - 33% on
desktop and 50% on mobile. Clear accordingly to avoid overlapping
infos.

Note that the original issue deals with debts but this actually
targets all material item lists.