4497: Avoid token race conditions when loading React components
React components which work for authenticated users are dependent on
tokens being set before they work as expected. This current way
tokens are loaded in parallel with component mounting can lead to
problems if the components are mounted before the tokens are loaded.

The possibility of such race conditions is highlighted by changes to
the checklist material button which supports resuming adding a
material after returning from an external login. The action will run
immediatly after returning and thus the token must be loaded and
available before this for the action to succeed.

To ensure the right loading order we load the file with the user
tokens before we mount the components.