5193: Fix section panels template being overriden
Ding Sections are supports to create a new panels variant for each
section page where editors can decide what panes to use for the
specific term/page. This is controlled by the term page field which
is set by default.
 
In issue 5037 the weight for the section tab which places
the field in the vertical tabs section was set in order to move the
section to the bottom of the section form.

Unfortunately for some reason this did not result in moving the
section down. Instead the term page field is removed from the form
entirely. This means that is does not have a value. Due to the way
the module is constructed this will instead result in editing the
general panels variant for sections.

This makes it seem as if new sections copy the old ones instead of
starting from scratch. Working with new sections will also affect
other sections which also accidentally use the general variant.

To address this this commits introduces the following changes:

1. Remove the #weight setting causing problems,
2. Set field order for current sites. Since the ding_sections do not
use features we have to do this manually using an update hook
combining field instances and bundle settings,
3. Set field order for new sites mirroring the changes in 2
4. Revert the ding sections general panels variant to the default
setup.