5193: Show teaser as a part of the section term description
Section panels currently use the term description panel pane to 
display the term. However this pane will only output the default
description field of the term.

We also want to output the teaser between the title and the
description for sections.

This is complicated by the fact that the term description pane is
already used in multiple section variants.

To overcome this we alter the definition of the pane plugin to use
a custom render function which output the default term view instead.

By adding a template specifically for sections we can control which
fields are shown. This would otherwise have to be handled through
display weights which are difficult to handle programmatically 
without Features.  

A little SCSS ensures that the teaser is shown with the same styling
as teasers/leads elsewhere on the site.

This approach should be OK as the term description pane is not used
elsewhere on the site.