5193: Transfer section title to menu item on save
Currently the current section title is transferred to the menu item
title field when the section form is shown. This means that if the
editor updates the section title then menu item is not updated. The
editor has to load the form for the new title to be transferred to
the menu title field and save again to have the menu title updated.

This change transfers the content of the section title field to the 
menu item title field before saving to ensure that they stay in sync.