Core: Fix Behat tests for PHP 7.4
For some reason Behat/Chrome will not navigate to the first url in
the search result but stays on the current page.

This means that the test will compare actual search results on the
last page with the expected search results of the first page. These
do not match.

The tests then fail when it cannot find a link to page 2 of the
search result in the pagination which on the last page contains
links to the first page and second to last page etc.

I have not been able to locate the actual source of this problem.

To address it this change moves from clicking on links in the
pagination to locating the corresponding urls and moving the browser
to them directly.

Here the urls must be absolute so a method for converting relative
urls in the links to absolute is also introduced.