Core: Fix Drupal coding standard check on Scrutinizer
A process is described at
https://scrutinizer-ci.com/docs/tools/php/code-sniffer/#problems-with-drupal-coding-standard
but that description does not work for us. Problems regarding a
loading Sleevomat classes remain.

Bases on the version history for the drupal/coder package we can tell
that this dependency was introduced in version 8.3.14 of drupal/coder
which was recently released. Scrutinizer will install version ^8.3 by
default.

To address this we pin coder to the last known working version.

When using this approach third party packages will end up in
./vendor. We do not want to include this in our analysis so we
configure PHPCS to ignore them.